% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached data}
\usage{
clear_cache(bfc, hash)
}
\arguments{
\item{bfc}{A `BiocFileCache` object used to manage cached data.}

\item{hash}{A character string representing the unique hash for the cached data.}
}
\value{
Logical `TRUE` if cache entries were successfully removed.
}
\description{
Removes all cached entries associated with a specific hash from the BiocFileCache system.
Deletes the corresponding local cache files and removes their entries from the cache database.
}
\keyword{internal}
