% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{create_cache}
\alias{create_cache}
\title{Create a Cache Entry for API Data}
\usage{
create_cache(path, bfc, hash, result_data)
}
\arguments{
\item{path}{A character string specifying the directory where the cache should be stored.}

\item{bfc}{A `BiocFileCache` object used to manage cached data.}

\item{hash}{A character string representing the unique hash for the cached data.}

\item{result_data}{The API response data to be saved in the cache.}
}
\value{
None. The function saves the data in the cache and prints a message.
}
\description{
This function stores API response data in a cache directory using `BiocFileCache`.
A temporary file is created in the specified cache location, and the data is saved in `.rds` format.
The cache entry is then registered in the `BiocFileCache` system to allow for easy retrieval later.
}
\keyword{internal}
