% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{create_hash}
\alias{create_hash}
\title{Generate a Unique Hash for API Caching}
\usage{
create_hash(endpoint, ...)
}
\arguments{
\item{endpoint}{A character string specifying the API endpoint (e.g., `"/lookup/symbol/"`).}

\item{...}{Additional arguments that are passed to the API endpoint function.}
}
\value{
A character string representing the unique MD5 hash.
}
\description{
This function generates a unique MD5 hash based on the API endpoint and
provided arguments. It ensures consistent hashing by sorting the arguments
before constructing the hash string.
}
\keyword{internal}
