% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_data_with_cache}
\alias{fetch_data_with_cache}
\title{Fetch Data with Caching Mechanism}
\usage{
fetch_data_with_cache(hash, fetch_function, ...)
}
\arguments{
\item{hash}{A unique identifier for caching the fetched data.}

\item{fetch_function}{A function that retrieves the required data when the cache is outdated or missing.}

\item{...}{Additional arguments to be passed to the fetch function.}
}
\value{
The fetched data, either from cache or a fresh API request.
}
\description{
This internal function handles fetching data from an API while implementing caching.
It first checks if the requested data is available in the cache.
If cached data is up to date, it returns the cached data; otherwise, it fetches new data,
updates the cache, and returns the latest results.
}
\keyword{internal}
