% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{get_info_assembly_species}
\alias{get_info_assembly_species}
\title{List the currently available assemblies for a species, along with toplevel sequences, chromosomes and cytogenetic bands.}
\usage{
get_info_assembly_species(species, bands = NULL)
}
\arguments{
\item{species}{A character string of species name or alias.}

\item{bands}{An optional boolean (`0` or `1`). If set to `1`, include cytogenetic band information.
Only display if band information is available.}
}
\value{
A list containing assembly-related information retrieved from the Ensembl REST API.
If the API request fails, an error message with an HTTP status code is returned.
}
\description{
Sends a GET request to the Ensembl REST API to retrieve assembly-related information for speciefic species.
}
\keyword{internal}
