% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{get_info_genomes_division_division_name}
\alias{get_info_genomes_division_division_name}
\title{Find information about all genomes in a given division.}
\usage{
get_info_genomes_division_division_name(division)
}
\arguments{
\item{division}{A character string specifying the Ensembl division (e.g., `"EnsemblVertebrates"`).}
}
\value{
A list containing genome-related information for the given division from the Ensembl REST API.
}
\description{
Fetches genome-related information for a specified division using the Ensembl REST API.
}
\keyword{internal}
