% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{get_info_species}
\alias{get_info_species}
\title{Retrieve Species Information from Ensembl REST API}
\usage{
get_info_species(division, ...)
}
\arguments{
\item{division}{A character string specifying the Ensembl division (e.g., `"EnsemblVertebrates"`).}

\item{...}{Named arguments representing optional parameters for the API request.}
}
\value{
A list of species data retrieved from the Ensembl REST API.
}
\description{
Sends a GET request to the Ensembl REST API to retrieve species-related information for speciefic division.
}
\keyword{internal}
