% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_request}
\alias{get_request}
\title{Perform a GET Request to an API}
\usage{
get_request(url)
}
\arguments{
\item{url}{A character string specifying the API endpoint.}
}
\value{
A list containing either the parsed JSON response or an error message with an HTTP status code.
}
\description{
Sends a GET request to the specified URL and retrieves data in JSON format and returns a structured response.
If the request fails, an error code and message are returned instead.
}
\keyword{internal}
