% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{info_assembly_species}
\alias{info_assembly_species}
\title{List the currently available assemblies for a species, along with toplevel sequences, chromosomes and cytogenetic bands.}
\usage{
info_assembly_species(species, bands = NULL)
}
\arguments{
\item{species}{A character string of species name or alias.}

\item{bands}{An optional boolean (`0` or `1`). If set to `1`, include cytogenetic band information.
Only display if band information is available.}
}
\value{
A list containing assembly-related information retrieved from the Ensembl REST API.
If the API request fails, an error message with an HTTP status code is returned.
}
\description{
Fetches assembly-related information for a specified species using the Ensembl REST API.
The function retrieves available genome assemblies, including toplevel sequences, chromosomes,
and optional cytogenetic band information.
}
\examples{
# Retrieve assembly information for anopheles_melas
info_assembly_species(species = "anopheles_melas")

# Retrieve assembly information for homo_sapiens including cytogenetic band information
info_assembly_species(species = "human", bands = 1)
}
\seealso{
\url{https://rest.ensembl.org/documentation/info/assembly_info}
}
