% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{info_genomes_genome_name}
\alias{info_genomes_genome_name}
\title{Find genome information for given genome name.}
\usage{
info_genomes_genome_name(name)
}
\arguments{
\item{name}{A character string of the genome name.}
}
\value{
A list containing genome-related information from the Ensembl REST API.
If the API request fails, an error message with an HTTP status code is returned.
}
\description{
Fetches genome-related information for a given genome name using the Ensembl REST API.
}
\examples{
# Retrieve genome information for arabidopsis_thaliana
info_genomes_genome_name(name = "arabidopsis_thaliana")
}
\seealso{
\url{https://rest.ensembl.org/documentation/info/info_genome}
}
