% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{info_karyotype_band}
\alias{info_karyotype_band}
\title{Find information about available cytogenetic band information for a given species.}
\usage{
info_karyotype_band(species)
}
\arguments{
\item{species}{A character string of species name or alias.}
}
\value{
A list containing cytogenetic band information retrieved from the Ensembl REST API.
}
\description{
Find information about available cytogenetic band information for a given species.
}
\examples{
# Retrieve karyotype band information for human.
info_karyotype_band(species = "human")
}
