% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{info_species}
\alias{info_species}
\title{Lists all available species, their aliases, available adaptor groups and data release.}
\usage{
info_species(division, ...)
}
\arguments{
\item{division}{A character string specifying the division. Accepted values:
`"Metazoa"`, `"Fungi"`, `"Plants"`, `"Protists"`, `"Vertebrates"`, or `"Bacteria"`.}

\item{...}{Named arguments representing optional parameters for the API request.}
}
\value{
A list containing species data from the Ensembl REST API.
}
\description{
Fetches species information from the Ensembl REST API for a given division.
}
\examples{
# Retrieve species data for Ensembl Plants
info_species(division = "Plants")

# Retrieve species data with additional parameters
info_species(division = "Vertebrates", strain_collection = "mouse")
}
\seealso{
\url{https://rest.ensembl.org/documentation/info/species}
}
