% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_id.R
\name{post_sequence_id}
\alias{post_sequence_id}
\title{Request multiple types of sequence by stable identifier.}
\usage{
post_sequence_id(id = NULL, ...)
}
\arguments{
\item{id}{A character vector representing the Ensembl stable ID.}

\item{...}{Named arguments representing optional parameters for the API request.}
}
\value{
A named list containing Biostrings objects (`DNAStringSet`, `AAStringSet`) for each requested ID.
}
\description{
This function sends a POST request to the Ensembl REST API to retrieve sequence data for given IDs.
It processes the API response, converts sequences into appropriate Biostrings objects.
}
\keyword{internal}
