% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{read_cache}
\alias{read_cache}
\title{Read cached data}
\usage{
read_cache(bfc, hash)
}
\arguments{
\item{bfc}{A `BiocFileCache` object used to manage cached data.}

\item{hash}{A character string representing the unique hash for the cached data.}
}
\value{
The cached data if available; otherwise, `NULL` if no valid cache exists.
}
\description{
Retrieves and loads cached data from the BiocFileCache system using a unique hash.
If the cache entry exists, it reads the stored `.rds` file and returns the data.
}
\keyword{internal}
