% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{update_cache}
\alias{update_cache}
\title{Update the cache entry for API data}
\usage{
update_cache(path, bfc, hash, result_data)
}
\arguments{
\item{path}{A character string specifying the directory where the cache should be stored.}

\item{bfc}{A `BiocFileCache` object used to manage cached data.}

\item{hash}{A character string representing the unique hash for the cached data.}

\item{result_data}{The API response data to be updated in the cache.}
}
\value{
None. The function updates the data in the cache and prints a message.
}
\description{
Updates an existing cache entry in BiocFileCache by replacing outdated data with newly fetched data.
It first saves the new data to a temporary file, deletes the old cached file if it exists,
and then updates the cache entry with the new file.
}
\keyword{internal}
