% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{validate_species_name}
\alias{validate_species_name}
\title{Validate Species Name Against Ensembl Species List}
\usage{
validate_species_name(species_name)
}
\arguments{
\item{species_name}{A character string representing the species name to validate.}
}
\value{
If the species name is valid, logical `TRUE` is returned. If not, an error is thrown with suggested
        closest matching species names.
}
\description{
This function validates a given species name by checking its existence in the Ensembl database.
Ensembl species names and aliases are fetched from the cache if available. Otherwise, they are retrieved from the API
and cached for future use.
If the species name is found, the function returns `TRUE`. If not, it suggests the closest matches
based on the Damerau-Levenshtein distance and stops execution with an error message.
}
\keyword{internal}
