% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_id.R
\name{archive_id}
\alias{archive_id}
\title{Retrieve the latest version for a set of identifiers}
\usage{
archive_id(id = NULL)
}
\arguments{
\item{id}{A character vector representing one or more Ensembl stable ID to be queried.}
}
\value{
A list containing the archive details from Ensembl REST API.
}
\description{
This function sends a request to the Ensembl REST API to retrieve information for given IDs.
}
\examples{
#Example for a single ID
archive_id( id = "ENSG00000157764")

#Example for multiple IDs
archive_id(id = c("ENSG00000157764","ENSG00000248378"))
}
\seealso{
\url{https://rest.ensembl.org/documentation/info/archive_id_get}
}
