% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{get_info_genomes_genome_name}
\alias{get_info_genomes_genome_name}
\title{Retrieve Genome Information from Ensembl REST API}
\usage{
get_info_genomes_genome_name(name = NULL)
}
\arguments{
\item{name}{A character string specifying the name of the genome.}
}
\value{
A list containing genome-related information retrieved from the Ensembl REST API.
}
\description{
Sends a GET request to the Ensembl REST API to retrieve specific genome information.
}
\keyword{internal}
