% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_request}
\alias{get_request}
\title{Perform a GET Request to an API}
\usage{
get_request(url)
}
\arguments{
\item{url}{A character string specifying the API endpoint.}
}
\value{
A structured response object parsed from JSON.
}
\description{
Sends a GET request to the specified URL and retrieves data in JSON format.
}
\keyword{internal}
