% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{info_genomes_genome_name}
\alias{info_genomes_genome_name}
\title{Retrieve genome information for given genome name.}
\usage{
info_genomes_genome_name(name = NULL)
}
\arguments{
\item{name}{A character string specifying the name of the genome.}
}
\value{
A list containing genome-related information retrieved from the Ensembl REST API.
}
\description{
Fetches genome-related information for a given genome name using the Ensembl REST API.
The function supports caching to avoid redundant API calls and improve performance.
}
\details{
More details at \url{https://rest.ensembl.org/documentation/info/info_genome}
}
\examples{
# Retrieve genome information for Arabidopsis thaliana
info_genomes_genome_name(name = "arabidopsis_thaliana")
}
