% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_symbol.R
\name{lookup_symbol}
\alias{lookup_symbol}
\title{Find the species and database for a set of symbols in a linked external database.}
\usage{
lookup_symbol(species = NULL, symbols = NULL)
}
\arguments{
\item{species}{A character string of species name}

\item{symbols}{A character vector of one or multiple symbols}
}
\value{
A list containing the lookup result from the Ensembl REST API.
}
\description{
The function constructs the API request, validates the input parameters, and returns the retrieved information as a list.
It expects the species and symbol(s) as input.
It first checks if the requested data is available in the cache.
If a valid cached entry exists, it returns the cached data.
If the cache is outdated or missing, it fetches new data, updates the cache, and returns the latest results.
}
\examples{
#Example for a single symbol
lookup_symbol(species = "homo_sapiens", symbols = "BRCA2")

#Example for multiple symbols
lookup_symbol(species = "homo_sapiens", symbols = c("BRCA2", "BRAF"))

More details at \url{https://rest.ensembl.org/documentation/info/symbol_lookup}
}
