% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_symbol.R
\name{lookup_symbol}
\alias{lookup_symbol}
\title{Find the species and database for a set of symbols in a linked external database.}
\usage{
lookup_symbol(species = NULL, symbols = NULL)
}
\arguments{
\item{species}{A character string of species name}

\item{symbols}{A character vector of one or multiple symbols}
}
\value{
A list containing the lookup result from the Ensembl REST API.
If caching is enabled, it may return previously stored results instead of making an API request.
}
\description{
The function expects the species and symbol(s) as input.
It constructs the API request, validates the input parameters, and returns the retrieved information as a list.
It first checks if the requested data is available in the cache.
If a valid cached entry exists, it returns the cached data.
If the cache is outdated or missing, it fetches new data, updates the cache, and returns the latest results.
}
\examples{
#Example for a single symbol
lookup_symbol("homo_sapiens","BRCA2")

#Example for multiple symbols
lookup_symbol("homo_sapiens", c("BRCA2", "BRAF"))
}
