% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_id.R
\name{post_archive_id}
\alias{post_archive_id}
\title{Retrieve the latest version for a set of identifiers}
\usage{
post_archive_id(id = NULL)
}
\arguments{
\item{id}{A character vector representing one or more Ensembl stable ID to be queried.}
}
\value{
A data frame containing the archive details from Ensembl REST API.
}
\description{
This function sends a POST request to the Ensembl REST API to retrieve information for given ID(s).
}
\keyword{internal}
