% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_symbol.R
\name{post_lookup_symbol}
\alias{post_lookup_symbol}
\title{Perform POST Request to Ensembl REST API for Symbol Lookup}
\usage{
post_lookup_symbol(species = NULL, symbols = NULL)
}
\arguments{
\item{species}{A character string specifying the species (e.g., "homo_sapiens"). This is a required parameter.}

\item{symbols}{A character vector of gene symbols to look up. This is passed as the body of the POST request.}
}
\value{
A parsed JSON response containing the lookup results for the specified symbols.
}
\description{
Sends a POST request to the Ensembl REST API to look up information for a set of symbols.
Separates the optional parameters from the whole parameters list.
Right now, no optional parameter is used in this function.
Constructs a query URL based on the mandatory and optional parameters and includes the symbols in the request body.
}
\keyword{internal}
