% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_symbol.R
\name{query_string_optional_params}
\alias{query_string_optional_params}
\title{Creates a string containing optional parameters}
\usage{
query_string_optional_params(optional_params)
}
\arguments{
\item{optional_params}{A named list of optional parameters.}
}
\value{
A character string containing the formatted query parameters or an empty string if no parameter is provided.
}
\description{
Creates a query string containing optional parameters in "key=value" format, separated by semicolons.
Removes any NULL values from the optional parameters before constructing the string.
}
\keyword{internal}
