% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{validate_genome_name}
\alias{validate_genome_name}
\title{Validate Genome Name Against Ensembl Genome Names List}
\usage{
validate_genome_name(genome_name)
}
\arguments{
\item{genome_name}{A character string representing the genome name to validate.}
}
\value{
If the genome name is valid, logical `TRUE` is returned. If not, an error is thrown with suggested
closest matching genome names.
}
\description{
This function validates a given genome name by checking its existence in the Ensembl database.
Ensembl genome names are fetched from the cache if available. Otherwise, they are retrieved from the API
and cached for future use.
If the genome name is found, the function returns `TRUE`. If not, it suggests the closest matches
based on the Damerau-Levenshtein distance and stops execution with an error message.
}
\keyword{internal}
